Moved functions around, added infy while loop for solving
